requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
tensorflow==2.14.0
googlemaps==4.10.0
pydantic==2.5.2
pytest==7.4.3
jupyter==1.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
flask==3.0.0
requests==2.31.0 